requirements.txt
dotenv
os
google.generativeai
pandas
plotly